src/main/java/org/syphr/lametrictime/api/local/model/Applications.java
/*
 * Copyright 2017 Gregory P. Moyer
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package org.syphr.lametrictime.api.local.model;

import java.util.List;

public class Applications
{
    private List<Application> list;

    public List<Application> getList()
    {
        return list;
    }

    public void setList(List<Application> list)
    {
        this.list = list;
    }

    public Applications withList(List<Application> list)
    {
        setList(list);
        return this;
    }

    @Override
    public int hashCode()
    {
        final int prime = 31;
        int result = 1;
        result = prime * result + ((list == null) ? 0 : list.hashCode());
        return result;
    }

    @Override
    public boolean equals(Object obj)
    {
        if (this == obj)
        {
            return true;
        }
        if (obj == null)
        {
            return false;
        }
        if (getClass() != obj.getClass())
        {
            return false;
        }
        Applications other = (Applications)obj;
        if (list == null)
        {
            if (other.list != null)
            {
                return false;
            }
        }
        else if (!list.equals(other.list))
        {
            return false;
        }
        return true;
    }

    @Override
    public String toString()
    {
        StringBuilder builder = new StringBuilder();
        builder.append("Applications [list=");
        builder.append(list);
        builder.append("]");
        return builder.toString();
    }
}
